+++ src/ASM/ASMs3Drecovery.C
@@ -370,7 +370,7 @@ Go::SplineVolume* ASMs3D::projectSolutionLeastSquare (const IntegrandBase& integ
     RealArray::const_iterator knotit = basis.begin();
     std::vector<double> tmp;
     tmp.reserve(nGauss*(basis.numCoefs()-basis.order()));
-    for (size_t i = 0; i<=(basis.numCoefs()-basis.order());i++)
+    for (int i = 0; i<=(basis.numCoefs()-basis.order());i++)
     {
       double d = knotit[i+basis.order()]-knotit[i+basis.order()-1];
       for (int j = 0; j < nGauss; j++)
